Don't round the returned last seen timestamp
It's not entirely clear why the rounding to a nearest day was here in the first place,
but it could lead to confusion for the client that uses this timestamp
for provenance of time for a batch of updates (e.g. in lightning-rapid-gossip-sync::update_network_graph_from_byte_stream).
For example:
1. Channel update U1 is applied at time D1_SEC + 1
2. Full gossip sync is requested
3. Channel update U1 is returned with `last_seen_timestamp` rounded to D1_SEC
4. Update is applied to client local state with D1_SEC timestamp (it
   actually is backdated to D1_SEC - 7 * SECS_IN_DAY, but that doesn't
   matter)
5. Channel update U2 is applied at time D1_SEC + 2
6. Gossip sync at timestamp D1_SEC (or D1_SEC + 1) is requested
7. Channel update U2 is returned with `last_seen_timestamp` rounded to D1_SEC
8. Client tries to apply U2 with timestamp D1_SEC and fails with "Update had same timestamp as last processed update" as previous update has the same rounded timestamp, which is not expected as this is actually a new update.

Additional issues arise when this timestamp is used to derive
`last_rapid_gossip_sync_timestamp` as this leads to duplicate update
messages that need to be ignored, but arguably it shouldn't be used in
this situation at all and instead we should use the actual current time
as seen by the Rapid Gossip Server Postgres instance.