Use the configured network rather than extracing from the net graph
When we were only generating snapshots that contained parts of the
network graph, extracting the chain hash from the network graph
announcements themselves was fine. However, now that we also
serialize empty snapshots as a "no changes since your last sync"
file, we need to initialize the chain_hash to the real network.

Luckily, `config::network()` already exists to make this easy, so
we simply switch to that.

Fixes #92