Include node details when the node is new
In cases where we only receive node details for the first time
after the last-seen timestamp, we were mistakenly not including the
node info at all in the snapshot. This was the most egregious for
initial syncs, where we'd include no node info at all.

Luckily the fix is simple, just include a full node info sync in
such cases.

Fixes #100