Move network graph persistence to its own tokio task
The DB persistence loop is responsible for taking new gossip off of
the persistence queue and writing it to Postgres. If it gets
stalled, the logic pushing gossip onto the event queue might also
stall. In the next commit, we'll move the gossip persistence logic
into a separate tokio runtime to avoid deadlocks, but in order for
that to work it has to be the case that the gossip persistence
task can never block waiting on a task on the main runtime, which
would be possible via the `NetworkGraph` mutexes. Thus, here, we
move the graph writing to its own task in the main runtime.